pyproject.toml
[project]
name = "biosound"
version = "0.1.0"
description = "Convert biodiversity observations into ambient, structurally legible MIDI timelines"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Biodiversity Soundscapes Team"}
]
keywords = ["biodiversity", "sonification", "midi", "ambient", "ecology"]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "pretty_midi>=0.2.10",
    "streamlit>=1.28.0",
    "soundfile>=0.12.0",
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
midi = [
    "mido>=1.3.0",
]

[project.scripts]
biosound = "biosound.cli:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
